feat(get-all-movies): client can map external movies to domain

+++ backend/src/test/java/fr/github/brice/movies/business/gateway/stub/InMemoryMovieRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class InMemoryMovieRepository implements MovieRepository {
 
-    private final HashMap<Integer, Movie> entities = new HashMap<>();
+    private final HashMap<Long, Movie> entities = new HashMap<>();
 
     @Override
     public Movie save(Movie movie) {
